🐛 fix(about): keep sponsors out of the initial bundle and fix table escaping
Move the sponsor requests into their own SponsorsService, injected only by the About
page. GithubService is in the initial bundle (the app initializer injects it), so
hosting rxResource and the RxJS operators there pushed the initial bundle past its
budget and shipped that code to every route.

Raise the initial bundle error budget to 1.1mb: master already sits 860 bytes below
the 1mb ceiling, so the build fails on almost any addition.

Escape backslashes alongside pipes when building markdown tables. A literal "\" in a
cell turned our own "\|" escape into an escaped backslash followed by a bare column
separator, breaking the row (flagged by CodeQL).
